requirements.txt
flask
python-dotenv
sqlalchemy
pyodbc
